Facilitate extra options res forward models
Make it possible for the user to include additional arguments for the
forward models defined in share/ert/res. This is specifically ment for
the user to include --ignore-errors, the only argument not already
handled.

The empty default argument is further stripped when running each script.
This will be handled differently when the forward model descriptive
files are rewritten.