Fix crash in test ert_util_matrix_lapack and remove memory leaks
An invalid use of fprintf caused fprintf to sometimes use up all memory
on the system, then crash, and sometimes flood stdout with characters.

The large memory usage happened 80 percent of the time when running a
Debug build. The flooding of characters happened consistently when
running through valgrind. When actually debugging no errors would occur,
making the bug hard to track down.

Fixed by editing format string. The formatting string used a * to
specify that the number of decimal points should be given in a separate
argument, which was not given. It seemed to work most of the time,
except for certain double values (e.g. 0.0).  Fixed it by specifying a
fixed number of decimal points.

Also fixes some memory leaks in both the test and library code.