Enable pytest to run from source/python instead of the build folder
Because several tests were dependent on compiled values, it was not
possible to run python tests from the source folder. One of these
tests were checking for the flow reservoir simulator, the way that
is checked has been changed. Removing the copy of python test files
to the build folder should also speed up the build process.
The test checking that all test files were copied to the build folder
is no longer valid. Because of `python/__init__.py` the tests were
not running the source python, but the installed version.