Create Jenkins pipeline
Adding a Jenkinsfile and bash script to the repository. This helps
in creating a build that is easier to edit, review and reproduce.
The bash script is written to be as self contained as possible, so
that it can be run locally for testing. The Jenkinsfile is a thin
layer around this script.