Make gen_kw test run by changing key order
The order of the gen_kw keys have either changed or this test has
never passed. The order is not clear, as the order in the test does
not mirror the order in the config file. If the order matters this
could be a bug. Changed the test to check that the correct keys were
read, and refactored.