Add new observation scaling job
Adding a new observation scaling job where the logic has been moved
from the C-code to python. This job will allow the user to give a
different set of keys for calculating the scaling factor and applying
the scaling factor. The scaling factor is based on PCA-analysis and
the user can provide a subset of points from which to calculate the scaling
factor. An index list can also be provided for each key where the calculated
scaling factor will be applied.