Stop shell jobs from using data_root
This is a major break in all shell jobs in libres. Prior to this commit
all shell jobs would interpret relative paths first relative to the
configuration file and only if that file did not exist, relative to the
runpath. From now on users will have to use <CONFIG_PATH> to pick up
files relative to the configuration file.